Backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-jose>=3.3.0
pydantic>=1.10.0
redis>=4.5.0
prometheus_client>=0.16.0